rename useCanvasSigner file
