+++ packages/hooks/src/index.ts
@@ -3,5 +3,5 @@ export { useCanvas } from "./useCanvas.js"
 export { useSession } from "./useSession.js"
 export { useRoute } from "./useRoute.js"
 export { useConnect, useSessionWallet, useDisconnect } from "./MultichainConnectContext.js"
-export { useCanvasSessionWallet } from "./useCanvasSigner.js"
+export { useCanvasSessionWallet } from "./useCanvasSessionWallet.js"
 export { MultichainConnect } from "./MultichainConnect.js"
